Update the rhel6 fb module to new registergui api
Use signal based error handling instead of the
multiple levels of overrided methods. In the process
also remove the RHEL5 firstboot support and the
firstboot_base.py module it required.

The custom screens that were included in rhsm_login
were just to alter the UI flow as firstboot needed
(and really, for the way this module doesn't fit the
normal firstboot module or moduleSet approach).

Ditch the rhel5 firstboot compat stuff as well.