Widget packing changes to handle wide screens.
The height of these screens is packed to the top of the screen,
but the widget can fill on resize. For these screens, they should
stay mostly the same when resized, except the entry boxes will
expand.

Note that if placed in a very wide dialog, the entry boxes will
fill and expand to be very wide. In those cases, the widgets
for register screens should be packed into a container widget
to constrain them to something reasonable.